+++ sn_logging/src/lib.rs
@@ -12,7 +12,7 @@ pub mod error;
 pub mod metrics;
 
 use self::error::{Error, Result};
-use std::fs;
+use std::fmt;
 use std::path::PathBuf;
 use tracing_appender::non_blocking::WorkerGuard;
 use tracing_core::{Event, Level, Subscriber};
@@ -29,6 +29,37 @@ use tracing_subscriber::{
     registry::LookupSpan,
     Layer, Registry,
 };
+
+#[derive(Debug, Clone)]
+pub enum LogOutputDest {
+    Stdout,
+    Path(PathBuf),
+}
+
+#[derive(Debug, Clone)]
+pub enum LogFormat {
+    Default,
+    Json,
+}
+
+pub fn parse_log_format(val: &str) -> Result<LogFormat> {
+    match val {
+        "default" => Ok(LogFormat::Default),
+        "json" => Ok(LogFormat::Json),
+        _ => Err(Error::LoggingConfigurationError(
+            "The only valid values for this argument are \"default\" or \"json\"".to_string(),
+        )),
+    }
+}
+
+impl fmt::Display for LogOutputDest {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            LogOutputDest::Stdout => write!(f, "stdout"),
+            LogOutputDest::Path(p) => write!(f, "{}", p.to_string_lossy()),
+        }
+    }
+}
 
 const ALL_SN_LOGS: &str = "all";
 
@@ -76,46 +107,44 @@ impl TracingLayers {
     fn fmt_layer(
         &mut self,
         default_logging_targets: Vec<(String, Level)>,
-        optional_log_dir: &Option<PathBuf>,
-        json_output: bool,
+        output_dest: LogOutputDest,
+        format: LogFormat,
     ) -> Result<()> {
-        let layer = if let Some(log_dir) = optional_log_dir {
-            if fs::remove_dir_all(log_dir).is_ok() {
-                println!("Removed old logs from directory: {log_dir:?}");
-            }
-
-            println!("Logging to directory: {log_dir:?}");
-
-            let logs_max_lines = 5000;
-            let logs_uncompressed = 100;
-            let logs_max_files = 1000;
-
-            let (file_rotation, worker_guard) =
-                appender::file_rotater(log_dir, logs_max_lines, logs_uncompressed, logs_max_files);
-            self.guard = Some(worker_guard);
-
-            if json_output {
-                tracing_fmt::layer()
-                    .json()
-                    .flatten_event(true)
-                    .with_writer(file_rotation)
-                    .boxed()
-            } else {
+        let layer = match output_dest {
+            LogOutputDest::Stdout => {
+                println!("Logging to stdout");
                 tracing_fmt::layer()
                     .with_ansi(false)
-                    .with_writer(file_rotation)
+                    .with_target(false)
                     .event_format(LogFormatter::default())
                     .boxed()
             }
-        } else {
-            println!("Logging to stdout");
-            tracing_fmt::layer()
-                .with_ansi(false)
-                .with_target(false)
-                .event_format(LogFormatter::default())
-                .boxed()
-        };
+            LogOutputDest::Path(ref path) => {
+                std::fs::create_dir_all(path)?;
+                println!("Logging to directory: {path:?}");
 
+                let logs_max_lines = 5000;
+                let logs_uncompressed = 100;
+                let logs_max_files = 1000;
+
+                let (file_rotation, worker_guard) =
+                    appender::file_rotater(path, logs_max_lines, logs_uncompressed, logs_max_files);
+                self.guard = Some(worker_guard);
+
+                match format {
+                    LogFormat::Json => tracing_fmt::layer()
+                        .json()
+                        .flatten_event(true)
+                        .with_writer(file_rotation)
+                        .boxed(),
+                    LogFormat::Default => tracing_fmt::layer()
+                        .with_ansi(false)
+                        .with_writer(file_rotation)
+                        .event_format(LogFormatter::default())
+                        .boxed(),
+                }
+            }
+        };
         let targets = match std::env::var("SN_LOG") {
             Ok(sn_log_val) => {
                 println!("Using SN_LOG={sn_log_val}");
@@ -185,17 +214,17 @@ impl TracingLayers {
 /// Logging should be instantiated only once.
 pub fn init_logging(
     default_logging_targets: Vec<(String, Level)>,
-    log_dir: &Option<PathBuf>,
-    json_output: bool,
+    output_dest: LogOutputDest,
+    format: LogFormat,
 ) -> Result<Option<WorkerGuard>> {
     let mut layers = TracingLayers::default();
 
     #[cfg(not(feature = "otlp"))]
-    layers.fmt_layer(default_logging_targets, log_dir, json_output)?;
+    layers.fmt_layer(default_logging_targets, output_dest, format)?;
 
     #[cfg(feature = "otlp")]
     {
-        layers.fmt_layer(default_logging_targets.clone(), log_dir, json_output)?;
+        layers.fmt_layer(default_logging_targets.clone(), output_dest, format)?;
 
         match std::env::var("OTEL_EXPORTER_OTLP_ENDPOINT") {
             Ok(_) => layers.otlp_layer(default_logging_targets)?,