+++ sn_node/tests/data_with_churn.rs
@@ -9,10 +9,10 @@
 use safenode_proto::{safe_node_client::SafeNodeClient, NodeInfoRequest, RestartRequest};
 
 use bytes::Bytes;
-use eyre::{bail, Result};
+use eyre::{bail, eyre, Result};
 use rand::{rngs::OsRng, Rng};
 use sn_client::{Client, Error, Files};
-use sn_logging::init_logging;
+use sn_logging::{init_logging, LogFormat, LogOutputDest};
 use sn_peers_acquisition::parse_peer_addr;
 use sn_protocol::{
     storage::{ChunkAddress, RegisterAddress},
@@ -110,8 +110,11 @@ async fn data_availability_during_churn() -> Result<()> {
         ("sn_networking".to_string(), Level::TRACE),
         ("sn_node".to_string(), Level::TRACE),
     ];
-    let log_appender_guard =
-        init_logging(logging_targets, &Some(tmp_dir.join("safe-client")), false)?;
+    let log_appender_guard = init_logging(
+        logging_targets,
+        LogOutputDest::Path(tmp_dir.join("safe-client")),
+        LogFormat::Default,
+    )?;
 
     println!("Creating a client...");
     let client = get_client().await;
@@ -439,16 +442,19 @@ async fn node_restart(addr: SocketAddr) -> Result<()> {
 
     let response = client.node_info(Request::new(NodeInfoRequest {})).await?;
     let log_dir = Path::new(&response.get_ref().log_dir);
+    let root_dir = log_dir
+        .parent()
+        .ok_or_else(|| eyre!("could not obtain parent from logging directory"))?;
 
     // remove Chunks records
-    let chunks_records = log_dir.join("record_store");
+    let chunks_records = root_dir.join("record_store");
     if let Ok(true) = chunks_records.try_exists() {
         println!("Removing Chunks records from {}", chunks_records.display());
         remove_dir_all(chunks_records).await?;
     }
 
     // remove Registers records
-    let registers_records = log_dir.join("registers");
+    let registers_records = root_dir.join("registers");
     if let Ok(true) = registers_records.try_exists() {
         println!(
             "Removing Registers records from {}",